+++ models/cnn_model.py
